modify the purchase and restock logic in service
